requirements.txt
Django>=5.1.7
djangorestframework>=3.15
django-phonenumber-field>=8.0.0
phonenumbers>=8.13
Pillow>=10.0
djangorestframework-simplejwt>=5.3
django-filter>=24.1
drf-spectacular>=0.27
